--- src/wsngamerz/mcmmo-PE/skills/SkillManager.php
@@ -1,34 +1,0 @@
-<?php
-namespace mcmmo-PE/skills;
-
-use mcmmo-PE\skills\acrobatics\Acrobatics;
-use mcmmo-PE\skills\acrobatics\AcrobaticsManager;
-use mcmmo-PE\skills\archery\Archery;
-use mcmmo-PE\skills\archery\ArcheryManager;
-use mcmmo-PE\skills\archery\TrackedEntity;
-use mcmmo-PE\skills\axes\Axes;
-use mcmmo-PE\skills\axes\AxesManager;
-use mcmmo-PE\skills\child\ChildConfig;
-use mcmmo-PE\skills\child\FamilyTree;
-use mcmmo-PE\skills\excavation\Excavation;
-use mcmmo-PE\skills\excavation\ExcavationManager;
-use mcmmo-PE\skills\herbalism\Herbalism;
-use mcmmo-PE\skills\herbalism\HerbalismManager;
-use mcmmo-PE\skills\repair\Repair;
-use mcmmo-PE\skills\repair\RepairManager;
-use mcmmo-PE\skills\salvage\Salvage;
-use mcmmo-PE\skills\salvage\SalvageManager;
-use mcmmo-PE\skills\smelting\Smelting;
-use mcmmo-PE\skills\smelting\SmeltingManager;
-use mcmmo-PE\skills\swords\Swords;
-use mcmmo-PE\skills\swords\SwordsManager;
-use mcmmo-PE\skills\taming\Taming;
-use mcmmo-PE\skills\taming\TamingManager;
-use mcmmo-PE\skills\unarmed\Unarmed;
-use mcmmo-PE\skills\unarmed\UnarmedManager;
-use mcmmo-PE\skills\woodcutting\Woodcutting;
-use mcmmo-PE\skills\woodcutting\WoodcuttingManager;
-
-public class SkillManager(
-  //this will be added to later
-}
