+++ src/wsngamerz/mcmmoPE/api/exceptions/InvalidSkillException.php
