+++ src/wsngamerz/mcmmoPE/skills/SkillManager.php
@@ -1,0 +1,34 @@
+<?php
+namespace mcmmoPE/skills;
+
+use mcmmoPE\skills\acrobatics\Acrobatics;
+use mcmmoPE\skills\acrobatics\AcrobaticsManager;
+use mcmmoPE\skills\archery\Archery;
+use mcmmoPE\skills\archery\ArcheryManager;
+use mcmmoPE\skills\archery\TrackedEntity;
+use mcmmoPE\skills\axes\Axes;
+use mcmmoPE\skills\axes\AxesManager;
+use mcmmoPE\skills\child\ChildConfig;
+use mcmmoPE\skills\child\FamilyTree;
+use mcmmoPE\skills\excavation\Excavation;
+use mcmmoPE\skills\excavation\ExcavationManager;
+use mcmmoPE\skills\herbalism\Herbalism;
+use mcmmoPE\skills\herbalism\HerbalismManager;
+use mcmmoPE\skills\repair\Repair;
+use mcmmoPE\skills\repair\RepairManager;
+use mcmmoPE\skills\salvage\Salvage;
+use mcmmoPE\skills\salvage\SalvageManager;
+use mcmmoPE\skills\smelting\Smelting;
+use mcmmoPE\skills\smelting\SmeltingManager;
+use mcmmoPE\skills\swords\Swords;
+use mcmmoPE\skills\swords\SwordsManager;
+use mcmmoPE\skills\taming\Taming;
+use mcmmoPE\skills\taming\TamingManager;
+use mcmmoPE\skills\unarmed\Unarmed;
+use mcmmoPE\skills\unarmed\UnarmedManager;
+use mcmmoPE\skills\woodcutting\Woodcutting;
+use mcmmoPE\skills\woodcutting\WoodcuttingManager;
+
+public class SkillManager{
+  //this will be added to later
+}
